auth: fix build on go1 and 386
In Go1, all integers were 32 bits. Now in Go 1.1/tip, they're 64 bits on
amd64, and 32 bits on 386. This commit fixes the build on travis (that
runs go 1.0.3) and 32 bits systems.

+++ auth/user.go
@@ -37,7 +37,7 @@ func loadConfig() error {
 	if cost == 0 && tokenExpire == 0 {
 		var err error
 		if days, err := config.GetInt("auth:token-expire-days"); err == nil {
-			tokenExpire = time.Duration(days * 24 * int(time.Hour))
+			tokenExpire = time.Duration(int64(days) * 24 * int64(time.Hour))
 		} else {
 			tokenExpire = defaultExpiration
 		}
